+++ breakwater-parser/src/lib.rs
@@ -1,53 +1,50 @@
 // Needed for simple implementation
 #![feature(portable_simd)]
 
-use enum_dispatch::enum_dispatch;
+use const_format::formatcp;
+
+mod assembler;
+mod framebuffer;
+mod memchr;
+mod original;
+mod refactored;
 
 #[cfg(target_arch = "x86_64")]
-pub mod assembler;
-pub mod memchr;
-pub mod original;
-pub mod refactored;
+pub use assembler::AssemblerParser;
+pub use framebuffer::{simple::SimpleFrameBuffer, FrameBuffer};
+pub use memchr::MemchrParser;
+pub use original::OriginalParser;
+pub use refactored::RefactoredParser;
 
-#[enum_dispatch(ParserImplementation)]
+pub const HELP_TEXT: &[u8] = formatcp!("\
+Pixelflut server powered by breakwater https://github.com/sbernauer/breakwater
+Available commands:
+HELP: Show this help
+PX x y rrggbb: Color the pixel (x,y) with the given hexadecimal color rrggbb
+{}
+PX x y gg: Color the pixel (x,y) with the hexadecimal color gggggg. Basically this is the same as the other commands, but is a more efficient way of filling white, black or gray areas
+PX x y: Get the color value of the pixel (x,y)
+{}SIZE: Get the size of the drawing surface, e.g. `SIZE 1920 1080`
+OFFSET x y: Apply offset (x,y) to all further pixel draws on this connection. This can e.g. be used to pre-calculate an image/animation and simply use the OFFSET command to move it around the screen without the need to re-calculate it
+",
+if cfg!(feature = "alpha") {
+    "PX x y rrggbbaa: Color the pixel (x,y) with the given hexadecimal color rrggbb and a transparency of aa, where ff means draw normally on top of the existing pixel and 00 means fully transparent (no change at all)"
+} else {
+    "PX x y rrggbbaa: Color the pixel (x,y) with the given hexadecimal color rrggbb. The alpha part is discarded for performance reasons, as breakwater was compiled without the alpha feature"
+},
+if cfg!(feature = "binary-commands") {
+    "PBxxyyrgba: Binary version of the PX command. x and y are little-endian 16 bit coordinates, r, g, b and a are a byte each. There is *no* newline after the command.\n"
+} else {
+    ""
+}
+).as_bytes();
+
+pub const ALT_HELP_TEXT: &[u8] = b"Stop spamming HELP!\n";
+
 pub trait Parser {
     /// Returns the last byte parsed. The next parsing loop will again contain all data that was not parsed.
     fn parse(&mut self, buffer: &[u8], response: &mut Vec<u8>) -> usize;
 
     // Sadly this cant be const (yet?) (https://github.com/rust-lang/rust/issues/71971 and https://github.com/rust-lang/rfcs/pull/2632)
     fn parser_lookahead(&self) -> usize;
-}
-
-#[enum_dispatch]
-pub enum ParserImplementation<FB: FrameBuffer> {
-    Original(original::OriginalParser<FB>),
-    Refactored(refactored::RefactoredParser<FB>),
-    Naive(memchr::MemchrParser<FB>),
-    #[cfg(target_arch = "x86_64")]
-    Assembler(assembler::AssemblerParser<FB>),
-}
-
-pub trait FrameBuffer {
-    fn get_width(&self) -> usize;
-
-    fn get_height(&self) -> usize;
-
-    fn get_size(&self) -> usize {
-        self.get_width() * self.get_height()
-    }
-
-    #[inline]
-    fn get(&self, x: usize, y: usize) -> Option<u32> {
-        if x < self.get_width() && y < self.get_height() {
-            Some(unsafe { self.get_unchecked(x, y) })
-        } else {
-            None
-        }
-    }
-
-    /// # Safety
-    /// make sure x and y are in bounds
-    unsafe fn get_unchecked(&self, x: usize, y: usize) -> u32;
-
-    fn set(&self, x: usize, y: usize, rgba: u32);
 }