+++ breakwater-parser/src/memchr.rs
@@ -2,20 +2,13 @@ use std::sync::Arc;
 
 use crate::{FrameBuffer, Parser};
 
-#[allow(dead_code)]
 pub struct MemchrParser<FB: FrameBuffer> {
-    help_text: &'static [u8],
-    alt_help_text: &'static [u8],
     fb: Arc<FB>,
 }
 
 impl<FB: FrameBuffer> MemchrParser<FB> {
-    pub fn new(fb: Arc<FB>, help_text: &'static [u8], alt_help_text: &'static [u8]) -> Self {
-        Self {
-            fb,
-            help_text,
-            alt_help_text,
-        }
+    pub fn new(fb: Arc<FB>) -> Self {
+        Self { fb }
     }
 }